refactoring some common learning code out
