+++ api/ellandi/reporting.py
@@ -7,6 +7,7 @@ from rest_framework import decorators, permissions, renderers, status
 from rest_framework.response import Response
 from rest_framework_csv.renderers import CSVRenderer
 
+from ellandi import learning
 from ellandi.registration.exceptions import MissingLanguageTypeError
 from ellandi.registration.models import (
     Grade,
@@ -55,19 +56,9 @@ SKILL_LEVELS = ["Beginner", "Advanced beginner", "Competent", "Proficient", "Exp
 LANGUAGE_LEVELS_SKILLED = ["Basic", "Independent", "Proficient", "Native"]
 LANGUAGE_TYPES = ["speaking", "writing"]
 
-HOURS_IN_WORK_DAY = 7.4
-LEARNING_TARGET_IN_DAYS = 10
-MINUTES_IN_LEARNING_TARGET = LEARNING_TARGET_IN_DAYS * HOURS_IN_WORK_DAY * 60
-
 
 def format_perc_label(number, percentage):
     return f"{number} ({round(percentage)}%)"
-
-
-def none_to_zero(number):
-    if not number:
-        number = 0
-    return number
 
 
 def create_proportions_data_dict(name, numerator, denominator):
@@ -437,80 +428,18 @@ def staff_overview_view(request):
     return Response(data=data, status=status.HTTP_200_OK, content_type="text/csv")
 
 
-def get_start_financial_year():
-    today = datetime.date.today()
-    year = today.year
-    if today.month < 4:
-        year = year - 1
-    start_financial_year = datetime.date(year, 4, 1)
-    return start_financial_year
-
-
-def get_learning_for_users_since_start_year(users_qs):
-    start_financial_year = get_start_financial_year()
-    learning_qs = Learning.objects.filter(user__in=users_qs).filter(date_completed__gte=start_financial_year)
-    return learning_qs
-
-
-def get_summary_course_costs(learning_qs):
-    formal_learning_with_costs = learning_qs.filter(learning_type="Formal").exclude(cost_unknown=True)
-    cost_aggregates = formal_learning_with_costs.aggregate(Sum("cost_pounds"), Avg("cost_pounds"))
-    avg_cost = none_to_zero(cost_aggregates["cost_pounds__avg"])
-    sum_cost = none_to_zero(cost_aggregates["cost_pounds__sum"])
-    return round(avg_cost), round(sum_cost)
-
-
-def get_percentage_for_learning_type(type, learning_qs, total_all_types):
-    if total_all_types == 0:
-        return 0
-    total_for_type = learning_qs.filter(learning_type=type).aggregate(Sum("duration_minutes"))["duration_minutes__sum"]
-    total_for_type = none_to_zero(total_for_type)
-    perc = total_for_type * 100 / total_all_types
-    return round(perc)
-
-
-def get_learning_distribution(learning_qs):
-    total_all_types = learning_qs.aggregate(Sum("duration_minutes"))["duration_minutes__sum"]
-    total_all_types = none_to_zero(total_all_types)
-    output = [
-        {
-            "name": "Formal",
-            "value_percentage": get_percentage_for_learning_type("Formal", learning_qs, total_all_types),
-        },
-        {
-            "name": "Social",
-            "value_percentage": get_percentage_for_learning_type("Social", learning_qs, total_all_types),
-        },
-        {
-            "name": "On the job",
-            "value_percentage": get_percentage_for_learning_type("On the job", learning_qs, total_all_types),
-        },
-    ]
-    return output
-
-
-def get_total_avg_learning_financial_year(learning_qs, total_users):
-    if not total_users:
-        return 0, 0
-    total_learning = learning_qs.aggregate(Sum("duration_minutes"))["duration_minutes__sum"]
-    total_learning = none_to_zero(total_learning)
-    avg_learning_mins = total_learning / total_users
-    proportion_learning_per_user = avg_learning_mins * 100 / MINUTES_IN_LEARNING_TARGET
-    days_per_user = round(avg_learning_mins / (60 * HOURS_IN_WORK_DAY))
-    avg_perc = round(proportion_learning_per_user)
-    return days_per_user, avg_perc
-
-
 @extend_schema(request=None, responses=None)
 @decorators.api_view(["GET"])
 @decorators.permission_classes((permissions.IsAdminUser,))
 def learning_view(request):
     users_qs = get_filtered_users(request)
     total_users = users_qs.count()
-    learning_qs = get_learning_for_users_since_start_year(users_qs)
-    course_average_cost, course_total_cost = get_summary_course_costs(learning_qs)
-    avg_completed_this_year, perc_completed_this_year = get_total_avg_learning_financial_year(learning_qs, total_users)
-    learning_distribution = get_learning_distribution(learning_qs)
+    learning_qs = learning.get_learning_for_users_since_start_year(users_qs)
+    course_average_cost, course_total_cost = learning.get_summary_course_costs(learning_qs)
+    avg_completed_this_year, perc_completed_this_year = learning.get_total_avg_learning_financial_year(
+        learning_qs, total_users
+    )
+    learning_distribution = learning.get_learning_distribution(learning_qs)
     output_dict = {
         "course_average_cost_label": f"£{course_average_cost:,}",
         "course_total_cost_label": f"£{course_total_cost:,}",
